split event into getter and parser
